requirements.txt
# CrewAI and Core Dependencies
crewai>=0.74.0
crewai-tools>=0.12.0
crewai[tools]

# Language Models
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0

# Search Tools
duckduckgo-search>=6.0.0
tavily-python>=0.5.0

# Streamlit and UI
streamlit>=1.28.0
streamlit-chat>=0.1.1

# OCR and Image Processing
easyocr>=1.7.0
Pillow>=10.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Additional Utilities
asyncio-throttle>=1.0.0
typing-extensions>=4.0.0

# Optional: For better performance
uvloop>=0.19.0; sys_platform != "win32"